Kernel+CL: Move to framebuffer and serial and format code.
This commit reworks the way logging is handled. Limine's terminal has
been ditched, being replaced with a framebuffer implementation.

A commonlib has appeared in the root directory of the project. This
contains code that should be available in both userspace and
kernel-space.

In the kernel, a new folder appeared: drivers. This folder contains
everything from video drivers (framebuffer, amd etc.) to communication
devices (serial, usb etc.). Additionaly, the framebuffer implementation
contains a hard-coded font file.

Since the serial device requires IO access, a new file called "io.h"
appeared in the "arch" directory. It implements wrappers around basic IO
instructions found in the x86 architecture (inb, outb etc.).

Besides those changes, additional formatting has been done as well as a
.clangd file has been provided for clangd users.

+++ kernel/interrupts/IDT.c
@@ -11,7 +11,7 @@ idt_set_descriptor(u8 vector, void *isr, u8 flags)
 
   vec->isr_low16 = addr & 0xFF;
   vec->isr_mid16 = (addr >> 16) & 0xFF;
-  vec->isr_high32 = addr >> 32;
+  vec->isr_high32 = (u32)(addr >> 32);
   vec->ist = 0;
   vec->zero = 0;
   vec->zero1 = 0;
